Refactor pub tests
