+++ tests/unit/transport/zeromq_test.py
@@ -28,6 +28,7 @@ from salt import client
 from salt.exceptions import EauthAuthenticationError, SaltInvocationError
 
 from unit.transport.req_test import ReqChannelMixin
+from unit.transport.pub_test import PubChannelMixin
 
 # TODO: move to a library?
 def get_config_file_path(filename):
@@ -159,38 +160,13 @@ class BaseZMQPubCase(AsyncTestCase):
         cls.req_server_channel.close()
 
 
-class AsyncPubChannelTest(BaseZMQPubCase):
+class AsyncPubChannelTest(BaseZMQPubCase, PubChannelMixin, TestCase):
     '''
     Tests around the publish system
     '''
     def get_new_ioloop(self):
         return  zmq.eventloop.ioloop.ZMQIOLoop()
-    def test_basic(self):
-        self.pub = None
-        def handle_pub(ret):
-            self.pub = ret
-            self.stop()
-        self.pub_channel = salt.transport.client.AsyncPubChannel.factory(self.minion_opts, io_loop=self.io_loop)
-        connect_future = self.pub_channel.connect()
-        connect_future.add_done_callback(lambda f: self.stop())
-        self.wait()
-        connect_future.result()
-        self.pub_channel.on_recv(handle_pub)
-        load = {
-                    'fun': 'f',
-                    'arg': 'a',
-                    'tgt': 't',
-                    'jid': 'j',
-                    'ret': 'r',
-                    'tgt_type': 'glob',
-                }
-        self.server_channel.publish(load)
-        self.wait()
-        self.assertEqual(self.pub['load'], load)
-        self.pub_channel.on_recv(None)
-        self.server_channel.publish(load)
-        with self.assertRaises(self.failureException):
-            self.wait(timeout=0.5)
+
 
 if __name__ == '__main__':
     from integration import run_tests
